requirements.txt
asgiref==3.4.1
astroid==2.3.3
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2
django-allauth==0.50.0
django-ckeditor==6.2.0
django-cleanup==5.2.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-richtextfield==1.6.1
django-tinymce==3.4.0
filelock==3.6.0
gunicorn==20.1.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==8.4.0
platformdirs==2.5.1
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.22.0
requests-oauthlib==1.3.1
six==1.13.0
sqlparse==0.4.2
sslcommerz==0.1
sslcommerz-lib==1.0
sslcommerz-python==0.0.7
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==20.14.0
whitenoise==5.3.0
wrapt==1.11.2
